Stop using `git stash`
No longer use `git stash` to save and restore the index.  Generate a
patch with the changes staged for the next commit using `git diff` and
restore it using `git apply` instead.

The problem with using `git stash` is that it creates a commit object
and all the associated files in the `.git` directory, so the size of
that directory will grow unnecessarily until garbage collection cleans
stuff up.